Removing mobileTopBar
